@@ -1,91 +1,66 @@
 {
-  "name": "page",
+  "name": "my-page",
+  "version": "1.0.0",
   "private": true,
+  "scripts": {
+    "dev": "next dev -p 8080",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
+    "export": "next build",
+    "serve": "next start",
+    "clean": "rm -rf .next out",
+    "test": "jest",
+    "updateSnap": "jest --updateSnapshot"
+  },
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/styles": "^4.11.3",
-    "@reach/router": "^1.3.4",
-    "axios": "^0.21.2",
-    "babel-plugin-prismjs": "^2.1.0",
-    "babel-preset-gatsby": "^1.1.0",
+    "@emotion/cache": "^11.14.0",
+    "@emotion/react": "^11.13.5",
+    "@emotion/styled": "^11.13.5",
+    "@mui/icons-material": "^6.1.9",
+    "@mui/material": "^6.5.0",
+    "@mui/material-nextjs": "^7.2.0",
+    "@mui/styles": "^6.5.0",
+    "axios": "^1.7.9",
     "bezier-easing": "^2.1.0",
-    "classnames": "^2.2.6",
-    "code-prettify": "^0.1.0",
-    "cross-env": "^7.0.3",
-    "event-source-polyfill": "^1.0.22",
-    "gatsby": "^3.15.0",
-    "gatsby-background-image": "^1",
-    "gatsby-cli": "^3",
-    "gatsby-image": "^3",
-    "gatsby-plugin-manifest": "^3",
-    "gatsby-plugin-material-ui": "^3",
-    "gatsby-plugin-netlify": "^3",
-    "gatsby-plugin-netlify-cms": "^5",
-    "gatsby-plugin-netlify-cms-paths": "^1.3.0",
-    "gatsby-plugin-offline": "^4",
-    "gatsby-plugin-react-helmet": "^4",
-    "gatsby-plugin-sharp": "^3",
-    "gatsby-plugin-styled-jsx": "^4",
-    "gatsby-remark-copy-linked-files": "^3",
-    "gatsby-remark-images": "^4",
-    "gatsby-source-filesystem": "^3",
-    "gatsby-transformer-remark": "^3",
-    "gatsby-transformer-sharp": "^3",
-    "github-markdown-css": "^3.0.1",
-    "gsap": "^3.6.0",
+    "classnames": "^2.5.1",
+    "github-markdown-css": "^5.8.1",
+    "gray-matter": "^4.0.3",
+    "gsap": "^3.12.8",
     "he": "^1.2.0",
-    "html-to-text": "^6.0.0",
-    "intersection-observer": "^0.12.0",
-    "markdown-to-txt": "^1.0.1",
-    "marked": "^4.0.10",
-    "material-ui-popup-state": "^1.8.0",
-    "netlify-cms-app": "^2.14.32",
-    "pixi.js": "^5.3.8",
-    "prismjs": "^1.27.0",
-    "prop-types": "^15.7.2",
-    "query-string": "^6.14.1",
-    "react": "^16.14.0",
-    "react-async": "^10.0.1",
-    "react-dom": "^16.14.0",
-    "react-helmet": "^6.1.0",
-    "react-paginate": "^7.1.0",
-    "scrollmagic": "^2.0.8",
-    "styled-jsx": "^3.4.4",
+    "html-to-text": "^9.0.5",
+    "intersection-observer": "^0.12.2",
+    "marked": "^15.0.2",
+    "material-ui-popup-state": "^5.1.2",
+    "next": "^15.1.0",
+    "pixi.js": "^8.11.0",
+    "prismjs": "^1.29.0",
+    "query-string": "^9.1.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-helmet-async": "^2.0.5",
+    "react-paginate": "^8.2.0",
     "typeface-roboto": "^1.1.13",
-    "uuid": "^3.4.0"
+    "uuid": "^11.0.3"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^24.9.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "eslint-plugin-jest": "^23.20.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^24.9.0",
-    "react-hot-loader": "^4.13.0",
-    "react-test-renderer": "^16.14.0"
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@types/node": "^22.10.2",
+    "@types/react": "^18.3.17",
+    "@types/react-dom": "^18.3.5",
+    "babel-plugin-prismjs": "^2.1.0",
+    "eslint": "^9.17.0",
+    "eslint-config-next": "^15.1.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "raw-loader": "^4.0.2",
+    "typescript": "^5.7.2"
   },
   "keywords": [
-    "gatsby"
+    "nextjs",
+    "portfolio",
+    "personal-website"
   ],
-  "license": "MIT",
-  "scripts": {
-    "cleanbuild": "gatsby clean && gatsby build",
-    "build": "cross-env GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES=true gatsby build --log-pages",
-    "develop": "gatsby develop -H 0.0.0.0 -p 8080",
-    "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
-    "start": "npm run develop",
-    "serve": "gatsby serve",
-    "test": "jest",
-    "clean": "gatsby clean",
-    "updateSnap": "jest --updateSnapshot",
-    "dev": "npm run develop"
-  }
+  "license": "MIT"
 }
