@@ -1,5 +1,5 @@
 import React from "react"
-import Paper from "@material-ui/core/Paper"
+import Paper from "@mui/material/Paper"
 
 // eslint-disable-next-line react/prop-types
 const Message = ({ children, color }) => {
