@@ -1,12 +1,13 @@
 /* eslint-disable react/jsx-props-no-spreading */
 /* eslint-disable react/prop-types */
+'use client'
+
 import React from "react"
 import PropTypes from "prop-types"
-import { makeStyles } from "@material-ui/styles"
+import { makeStyles } from "@mui/styles"
 
-const useStyles = makeStyles(height => ({
+const useStyles = makeStyles({
   root: {
-    height,
     width: "100%",
     userSelect: "none",
     WebkitUserSelect: "none",
@@ -14,17 +15,23 @@ const useStyles = makeStyles(height => ({
     // border: "red solid 2px",
     // pointerEvents: "none",
   },
-}))
+})
 
 // Section component whose parent should be type of Container
 const Section = React.forwardRef(({ children, style, className, id, height = "100vh" }, ref) => {
-  const classes = useStyles(height)
+  const classes = useStyles()
 
   return (
     <section
       ref={ref}
       className={`${classes.root} section ${className || ""}`}
-      style={{ boxSizing: "border-box", userSelect: "none", WebkitUserSelect: "none", ...style }}
+      style={{
+        height,
+        boxSizing: "border-box",
+        userSelect: "none",
+        WebkitUserSelect: "none",
+        ...style
+      }}
       {...(id ? { id } : {})}
     >
       {children}
