@@ -1,5 +1,5 @@
 import React from "react"
-import Grid from "@material-ui/core/Grid"
+import Grid from "@mui/material/Grid"
 
 // eslint-disable-next-line react/prop-types
 function CardContainer({ children, style }) {
