@@ -1,12 +1,14 @@
+'use client'
+
 /* eslint-disable react/prop-types */
 import React from "react"
-import { makeStyles } from "@material-ui/core/styles"
-import Card from "@material-ui/core/Card"
-import CardActionArea from "@material-ui/core/CardActionArea"
-import CardContent from "@material-ui/core/CardContent"
-import CardMedia from "@material-ui/core/CardMedia"
-import Typography from "@material-ui/core/Typography"
-import Grid from "@material-ui/core/Grid"
+import { makeStyles } from "@mui/styles"
+import Card from "@mui/material/Card"
+import CardActionArea from "@mui/material/CardActionArea"
+import CardContent from "@mui/material/CardContent"
+import CardMedia from "@mui/material/CardMedia"
+import Typography from "@mui/material/Typography"
+import Grid from "@mui/material/Grid"
 
 const useStyles = makeStyles({
   card: {
