@@ -1,35 +1,13 @@
 /* eslint-disable react/prop-types */
 import React from "react"
-import { makeStyles } from "@material-ui/core/styles"
-import Card from "@material-ui/core/Card"
-import CardActionArea from "@material-ui/core/CardActionArea"
-import CardActions from "@material-ui/core/CardActions"
-import CardContent from "@material-ui/core/CardContent"
-import CardMedia from "@material-ui/core/CardMedia"
-import Button from "@material-ui/core/Button"
-import Typography from "@material-ui/core/Typography"
-import Grid from "@material-ui/core/Grid"
-
-const useStyles = makeStyles({
-  card: {
-    maxWidth: "100%",
-    backgroundColor: "transparent",
-    height: "100%",
-  },
-  actionArea: {
-    height: "90%",
-  },
-  media: {
-    height: "65%",
-  },
-  content: {
-    height: "35%",
-    overflow: "hidden",
-  },
-  action: {
-    height: "10%",
-  },
-})
+import Card from "@mui/material/Card"
+import CardActionArea from "@mui/material/CardActionArea"
+import CardActions from "@mui/material/CardActions"
+import CardContent from "@mui/material/CardContent"
+import CardMedia from "@mui/material/CardMedia"
+import Button from "@mui/material/Button"
+import Typography from "@mui/material/Typography"
+import Grid from "@mui/material/Grid"
 
 export default function MediaCard({
   onClick,
@@ -38,14 +16,13 @@ export default function MediaCard({
   title,
   description,
 }) {
-  const classes = useStyles()
   return (
     <Grid item xs={5} style={{ height: "450px", maxHeight: "100%", ...style }}>
-      <Card className={classes.card}>
-        <CardActionArea onClick={onClick} className={classes.actionArea}>
+      <Card sx={{ maxWidth: "100%", backgroundColor: "transparent", height: "100%" }}>
+        <CardActionArea onClick={onClick} sx={{ height: "90%" }}>
           {image ? (
             <CardMedia
-              className={classes.media}
+              sx={{ height: "65%" }}
               image={
                 image.childImageSharp ? image.childImageSharp.fluid.src : image
               }
@@ -56,9 +33,8 @@ export default function MediaCard({
               container
               justifyContent="center"
               alignItems="center"
-              className={classes.media}
+              sx={{ height: "65%", backgroundColor: "white" }}
               title={title}
-              style={{ backgroundColor: "white" }}
             >
               <div style={{ margin: "5%" }}>
                 <h1>
@@ -67,7 +43,7 @@ export default function MediaCard({
               </div>
             </Grid>
           )}
-          <CardContent className={classes.content}>
+          <CardContent sx={{ height: "35%", overflow: "hidden" }}>
             <Typography
               gutterBottom
               variant="h6"
@@ -78,7 +54,7 @@ export default function MediaCard({
             </Typography>
             <Typography
               variant="body2"
-              color="textSecondary"
+              color="text.secondary"
               component="p"
               style={{ height: "50%", overflow: "hidden" }}
             >
@@ -86,7 +62,7 @@ export default function MediaCard({
             </Typography>
           </CardContent>
         </CardActionArea>
-        <CardActions className={classes.action}>
+        <CardActions sx={{ height: "10%" }}>
           <Button size="small" color="primary" onClick={onClick}>
             Learn More
           </Button>
