@@ -1,17 +1,19 @@
 import { useCallback, useLayoutEffect } from "react"
-import useRecordComponentState from "./useRestoreComponentState"
+import { useRestoreComponentStateToBeforeRouting } from "./useRestoreComponentState"
 import useLayoutContext from "./useLayoutContext"
 
 export default function(path, hash) {
+  path = [...path, "scroll"]
   const context = useLayoutContext()
   const getCurrentState = useCallback(() => {
     return {
-      scrollTop:
+      scroll_percent:
         context.scrollLayer.scrollTop / context.scrollLayer.scrollHeight,
     }
-  }, [...path])
+  }, [])
 
-  const historyState = useRecordComponentState(path, getCurrentState)
+  // use before-routing instead of unmounting to avoid reading scroll position of 0 during unmounting
+  const historyState = useRestoreComponentStateToBeforeRouting(path, getCurrentState)
   // restore to history state
   useLayoutEffect(() => {
     let scrollTop = 0
@@ -21,8 +23,8 @@ export default function(path, hash) {
     if (target) {
       scrollTop = target.offsetTop
     } else if (historyState) {
-      scrollTop = historyState.scrollTop * context.scrollLayer.scrollHeight
+      scrollTop = historyState.scroll_percent * context.scrollLayer.scrollHeight
     }
     context.scrollLayer.scrollTop = scrollTop
-  }, [])
+  }, [path, hash])
 }
